Abstracted PIN input into a component
